backend/requirements.txt
Django>=3.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
celery
redis
django-celery-beat
python-dotenv
dj-database-url
argon2-cffi  
psycopg[binary]>=3.2.10,<3.4
gunicorn==22.0.0